Eliminate type duplication: unify Attribute types
- Create shared src/attribute.rs module for Attribute type
- Remove duplicate Attribute definitions from tagger.rs and trainer.rs
- Update all imports to use unified crate::Attribute
- Update examples and tests to use single Attribute type
- Run cargo fmt to format all code

This addresses review feedback to eliminate duplication between
TaggerAttribute and training Attribute types.
